requirements.txt
streamlit>=1.31.0
openai>=1.11.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.26.0

# Testing requirements
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-mock>=3.11.1